+++ tests/fixtures.py
@@ -13,6 +13,7 @@ class FakeEnv:
     URL = 'http://www.google.com/'
     POSTGRES = 'postgres://uf07k1:wegauwhg@ec2-107-21-253-135.compute-1.amazonaws.com:5431/d8r82722'
     MYSQL = 'mysql://bea6eb0:69772142@us-cdbr-east.cleardb.com/heroku_97681?reconnect=true'
+    MYSQL_CLOUDSQL_URL = 'mysql://djuser:hidden-password@//cloudsql/arvore-codelab:us-central1:mysqlinstance/mydatabase'
     MYSQLGIS = 'mysqlgis://user:password@127.0.0.1/some_database'
     SQLITE = 'sqlite:////full/path/to/your/database/file.sqlite'
     ORACLE_TNS = 'oracle://user:password@sid/'
@@ -31,6 +32,8 @@ class FakeEnv:
     def generate_data(cls):
         return dict(STR_VAR='bar',
                     MULTILINE_STR_VAR='foo\\nbar',
+                    MULTILINE_QUOTED_STR_VAR='---BEGIN---\\r\\n---END---',
+                    MULTILINE_ESCAPED_STR_VAR='---BEGIN---\\\\n---END---',
                     INT_VAR='42',
                     FLOAT_VAR='33.3',
                     FLOAT_COMMA_VAR='33,3',
@@ -51,6 +54,7 @@ class FakeEnv:
                     BOOL_FALSE_STRING_LIKE_BOOL='False',
                     BOOL_FALSE_BOOL=False,
                     PROXIED_VAR='$STR_VAR',
+                    ESCAPED_VAR=r'\$baz',
                     INT_LIST='42,33',
                     INT_TUPLE='(42,33)',
                     STR_LIST_WITH_SPACES=' foo,  bar',
@@ -64,6 +68,7 @@ class FakeEnv:
                     DATABASE_ORACLE_TNS_URL=cls.ORACLE_TNS,
                     DATABASE_REDSHIFT_URL=cls.REDSHIFT,
                     DATABASE_CUSTOM_BACKEND_URL=cls.CUSTOM_BACKEND,
+                    DATABASE_MYSQL_CLOUDSQL_URL=cls.MYSQL_CLOUDSQL_URL,
                     CACHE_URL=cls.MEMCACHE,
                     CACHE_REDIS=cls.REDIS,
                     EMAIL_URL=cls.EMAIL,